# server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "IT Service Desk API",
    "description": "API documentation for IT Service Desk application - incident and equipment management system with JWT authentication, notifications, and analytics",
    "version": "1.0.0",
    "contact": {
      "name": "IT Service Desk Support",
      "email": "support@itservicedesk.example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "IncidentResponse": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Incident ID",
            "example": "60d21b4667d0d8992e610c85"
          },
          "title": {
            "type": "string",
            "description": "Incident title",
            "example": "Computer not starting"
          },
          "description": {
            "type": "string",
            "description": "Incident description",
            "example": "My computer won't turn on after power outage"
          },
          "priority": {
            "type": "string",
            "enum": ["low", "medium", "high", "critical"],
            "description": "Incident priority",
            "example": "high"
          },
          "status": {
            "type": "string",
            "enum": ["new", "assigned", "in_progress", "on_hold", "resolved", "closed"],
            "description": "Incident status",
            "example": "new"
          },
          "category": {
            "type": "string",
            "description": "Incident category",
            "example": "Hardware"
          },
          "reportedBy": {
            "$ref": "#/components/schemas/UserResponse"
          },
          "assignedTo": {
            "$ref": "#/components/schemas/UserResponse"
          },
          "equipment": {
            "$ref": "#/components/schemas/EquipmentResponse"
          },
          "comments": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string",
                  "description": "Comment text",
                  "example": "Checked the power supply, it's damaged"
                },
                "user": {
                  "$ref": "#/components/schemas/UserResponse"
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time",
                  "description": "Comment creation date",
                  "example": "2023-03-15T14:30:00Z"
                }
              }
            }
          },
          "resolutionNotes": {
            "type": "string",
            "description": "Notes on how the incident was resolved",
            "example": "Replaced power supply unit"
          },
          "resolutionDate": {
            "type": "string",
            "format": "date-time",
            "description": "Date when the incident was resolved",
            "example": "2023-03-16T10:15:00Z"
          },
          "attachments": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "filename": {
                  "type": "string",
                  "description": "Attachment filename",
                  "example": "error_screenshot.png"
                },
                "path": {
                  "type": "string",
                  "description": "Attachment file path",
                  "example": "/uploads/incidents/60d21b4667d0d8992e610c85/error_screenshot.png"
                },
                "mimetype": {
                  "type": "string",
                  "description": "Attachment MIME type",
                  "example": "image/png"
                },
                "size": {
                  "type": "number",
                  "description": "Attachment file size in bytes",
                  "example": 245678
                },
                "uploadedBy": {
                  "$ref": "#/components/schemas/UserResponse"
                },
                "uploadedAt": {
                  "type": "string",
                  "format": "date-time",
                  "description": "Attachment upload date",
                  "example": "2023-03-15T12:30:00Z"
                }
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Incident creation date",
            "example": "2023-03-15T12:00:00Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Incident last update date",
            "example": "2023-03-16T10:15:00Z"
          }
        }
      },
      "IncidentCreateRequest": {
        "type": "object",
        "required": ["title", "description", "priority", "category"],
        "properties": {
          "title": {
            "type": "string",
            "description": "Incident title",
            "example": "Computer not starting"
          },
          "description": {
            "type": "string",
            "description": "Incident description",
            "example": "My computer won't turn on after power outage"
          },
          "priority": {
            "type": "string",
            "enum": ["low", "medium", "high", "critical"],
            "description": "Incident priority",
            "example": "high"
          },
          "category": {
            "type": "string",
            "description": "Incident category",
            "example": "Hardware"
          },
          "equipment": {
            "type": "string",
            "description": "Equipment ID related to the incident",
            "example": "60d21b4667d0d8992e610c87"
          }
        }
      },
...
      "NotificationResponse": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Notification ID",
            "example": "60d21b4667d0d8992e610c88"
          },
          "title": {
            "type": "string",
            "description": "Notification title",
            "example": "New Incident Assigned"
          },
          "message": {
            "type": "string",
            "description": "Notification message",
            "example": "You have been assigned to incident #INC-2023-001"
          },
          "type": {
            "type": "string",
            "enum": ["info", "warning", "error", "success"],
            "description": "Notification type",
            "example": "info"
          },
          "recipient": {
            "$ref": "#/components/schemas/UserResponse"
          },
          "relatedEntity": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "enum": ["incident", "equipment", "user", "report"],
                "description": "Related entity type",
                "example": "incident"
              },
              "id": {
                "type": "string",
                "description": "Related entity ID",
                "example": "60d21b4667d0d8992e610c85"
              }
            }
          },
          "isRead": {
            "type": "boolean",
            "description": "Whether the notification has been read",
            "example": false
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Notification creation date",
            "example": "2023-03-15T13:00:00Z"
          },
          "readAt": {
            "type": "string",
            "format": "date-time",
            "description": "Date when the notification was read",
            "example": null
          }
        }
      },
...
      "PaginatedResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "description": "Array of items",
            "items": {
              "type": "object"
            }
          },
          "pagination": {
            "type": "object",
            "properties": {
              "total": {
                "type": "number",
                "description": "Total number of items",
                "example": 100
              },
              "limit": {
                "type": "number",
                "description": "Number of items per page",
                "example": 10
              },
              "page": {
                "type": "number",
                "description": "Current page number",
                "example": 1
              },
              "pages": {
                "type": "number",
                "description": "Total number of pages",
                "example": 10
              },
              "hasNext": {
                "type": "boolean",
                "description": "Whether there is a next page",
                "example": true
              },
              "hasPrev": {
                "type": "boolean",
                "description": "Whether there is a previous page",
                "example": false
              }
            }
          }
        }
      }
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Access token is missing or invalid",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "message": "Unauthorized",
              "statusCode": 401
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "User does not have permission to access this resource",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "message": "Forbidden",
              "statusCode": 403
            }
          }
        }
      },
...
      "ValidationError": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "example": {
              "message": "Validation failed",
              "errors": [
                {
                  "field": "email",
                  "message": "Email is required"
                },
                {
                  "field": "password",
                  "message": "Password must be at least 8 characters"
                }
              ],
              "statusCode": 400
            }
          }
        }
      },
      "ServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "message": "Internal Server Error",
              "statusCode": 500
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/incidents": {
      "get": {
        "tags": ["Incidents"],
        "summary": "Get all incidents",
        "description": "Get a paginated list of incidents with optional filtering",
        "operationId": "getIncidents",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by status",
            "schema": {
              "type": "string",
              "enum": ["new", "assigned", "in_progress", "on_hold", "resolved", "closed"]
            }
          },
          {
            "name": "priority",
            "in": "query",
            "description": "Filter by priority",
            "schema": {
              "type": "string",
              "enum": ["low", "medium", "high", "critical"]
            }
          },
          {
            "name": "category",
            "in": "query",
            "description": "Filter by category",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assignedTo",
            "in": "query",
            "description": "Filter by assigned technician ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reportedBy",
            "in": "query",
            "description": "Filter by reporter ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "equipment",
            "in": "query",
            "description": "Filter by equipment ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search in title and description",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "description": "Filter by creation date (start)",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "description": "Filter by creation date (end)",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sort field",
            "schema": {
              "type": "string",
              "enum": ["createdAt", "updatedAt", "priority", "status"],
              "default": "createdAt"
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "Sort order",
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"],
              "default": "desc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of incidents",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/IncidentResponse"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "post": {
        "tags": ["Incidents"],
        "summary": "Create incident",
        "description": "Create a new incident",
        "operationId": "createIncident",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Incident created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
    "/equipment": {
      "get": {
        "tags": ["Equipment"],
        "summary": "Get all equipment",
        "description": "Get a paginated list of equipment with optional filtering",
        "operationId": "getEquipment",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "Filter by equipment type",
            "schema": {
              "type": "string",
              "enum": ["laptop", "desktop", "server", "printer", "network", "peripheral", "other"]
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by equipment status",
            "schema": {
              "type": "string",
              "enum": ["available", "in_use", "under_repair", "reserved", "decommissioned"]
            }
          },
          {
            "name": "assignedTo",
            "in": "query",
            "description": "Filter by assigned user ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search in name, model, serial number",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "manufacturer",
            "in": "query",
            "description": "Filter by manufacturer",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "location",
            "in": "query",
            "description": "Filter by location",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "warrantyExpiring",
            "in": "query",
            "description": "Filter by warranty expiring in days",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sort field",
            "schema": {
              "type": "string",
              "enum": ["name", "type", "status", "purchaseDate", "warrantyExpiryDate"],
              "default": "name"
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "Sort order",
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"],
              "default": "asc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of equipment",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/EquipmentResponse"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "post": {
        "tags": ["Equipment"],
        "summary": "Create equipment",
        "description": "Create a new equipment item",
        "operationId": "createEquipment",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EquipmentCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Equipment created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EquipmentResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
    "/notifications": {
      "get": {
        "tags": ["Notifications"],
        "summary": "Get user notifications",
        "description": "Get a paginated list of notifications for the current user",
        "operationId": "getUserNotifications",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "isRead",
            "in": "query",
            "description": "Filter by read status",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "Filter by notification type",
            "schema": {
              "type": "string",
              "enum": ["info", "warning", "error", "success"]
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sort field",
            "schema": {
              "type": "string",
              "enum": ["createdAt", "readAt"],
              "default": "createdAt"
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "Sort order",
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"],
              "default": "desc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of notifications",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/NotificationResponse"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "post": {
        "tags": ["Notifications"],
        "summary": "Create notification",
        "description": "Create a new notification (admin and technician only)",
        "operationId": "createNotification",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Notification created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
    "/reports": {
      "get": {
        "tags": ["Reports"],
        "summary": "Get all reports",
        "description": "Get a paginated list of reports",
        "operationId": "getReports",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "Filter by report type",
            "schema": {
              "type": "string",
              "enum": ["incident", "equipment", "performance", "custom"]
            }
          },
          {
            "name": "format",
            "in": "query",
            "description": "Filter by report format",
            "schema": {
              "type": "string",
              "enum": ["pdf", "excel", "csv"]
            }
          },
          {
            "name": "createdBy",
            "in": "query",
            "description": "Filter by creator ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search in name and description",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "description": "Filter by creation date (start)",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "description": "Filter by creation date (end)",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sort field",
            "schema": {
              "type": "string",
              "enum": ["name", "type", "format", "createdAt"],
              "default": "createdAt"
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "Sort order",
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"],
              "default": "desc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of reports",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/ReportResponse"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "post": {
        "tags": ["Reports"],
        "summary": "Generate report",
        "description": "Generate a new report",
        "operationId": "generateReport",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Report generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}